refactor: 쿼리 파라미터 파싱 및 보관 RequestLine으로 책임 이동
